hardware-controller/requirements.txt
paho-mqtt
pytz
schedule
